refactor: migrate OptionStatus enum to a dedicated file - Consolidated the OptionStatus enum into a new enums/Options.ts file and updated all relevant components and tests to reference the new location, improving code organization and maintainability.
