+++ src/enums/Options.ts
@@ -1,0 +1,8 @@
+/**
+ * Enum para definir o status de uma alternativa/opção
+ */
+export enum OptionStatus {
+  CORRECT = 'correct',
+  INCORRECT = 'incorrect',
+  NEUTRAL = 'neutral',
+}
