+++ src/utils/questionRenderer/multipleChoice/index.tsx
@@ -1,9 +1,9 @@
 import type { ReactNode } from 'react';
 import { ANSWER_STATUS } from '../../../components/Quiz/useQuizStore';
 import { MultipleChoiceList } from '../../../components/MultipleChoice/MultipleChoice';
+import { OptionStatus } from '../../../enums/Options';
 import Text from '../../../components/Text/Text';
 import type { QuestionRendererProps } from '../types';
-import { Status } from '../types';
 
 /**
  * Render multiple choice question
@@ -32,18 +32,18 @@ export const renderQuestionMultipleChoice = ({
         result?.answerStatus !== ANSWER_STATUS.NAO_RESPONDIDO) ||
       hasAutoValidation;
 
-    let status: Status;
+    let status: OptionStatus;
     if (shouldShowCorrectAnswers) {
       if (isCorrectOption) {
-        status = Status.CORRECT;
+        status = OptionStatus.CORRECT;
       } else if (isSelected && !isCorrectOption) {
-        status = Status.INCORRECT;
+        status = OptionStatus.INCORRECT;
       } else {
-        status = Status.NEUTRAL;
+        status = OptionStatus.NEUTRAL;
       }
     } else {
       // When pending evaluation and no auto-validation, show all options as neutral
-      status = Status.NEUTRAL;
+      status = OptionStatus.NEUTRAL;
     }
 
     return {
