+++ src/utils/questionRenderer/types.ts
@@ -2,10 +2,6 @@ import type {
   Question,
   QuestionResult,
 } from '../../components/Quiz/useQuizStore';
-import { OptionStatus } from '../../components/Alternative/Alternative';
-
-// Re-export OptionStatus for backward compatibility
-export { OptionStatus as Status };
 
 /**
  * Props for question renderers
